pyproject.toml
[project]
name = "dapr-state-cache"
version = "0.5.0"
description = "Cache transparente para aplicações Dapr"
readme = "README.md"
authors = [
    { name = "Helton Dória", email = "helton.doria@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "msgpack>=1.1.1",
    "opentelemetry-api>=1.38.0",
]

[build-system]
requires = ["uv_build>=0.8.19,<0.10.0"]
build-backend = "uv_build"

[project.optional-dependencies]
dev = [
    "codespell>=2.4.1", # a tool for finding and fixing common misspellings in text files
    "mypy>=1.19.0", # a static type checker for Python
    "pyright>=1.1.407", # a static type checker for Python
    "pytest>=9.0.1", # a testing framework for Python
    "pytest-asyncio>=1.3.0", # a plugin for pytest that enables async test support
    "pytest-cov>=7.0.0", # a plugin for pytest that generates coverage reports
    "radon>=5.1.0", # a tool for measuring code complexity and performance
    "ruff>=0.14.7", # a fast, opinionated Python code formatter
    "vulture>=2.14", # open-source tool designed to find unused or "dead" code in Python 
    "wily>=1.25.0", # a tool for measuring code complexity and performance
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort (imports)
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "RUF", # regras do Ruff
]
ignore = [
    "UP047",  # type parameters (incompatível com ParamSpec em overloads)
    "RUF022", # __all__ sorting (preferimos agrupamento semântico)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ['S101']

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q -ra --strict-markers --disable-warnings --maxfail=1 --cov --cov-report=term-missing"
testpaths = ["tests"]
xfail_strict = true
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true
source = ["."]
omit = [
    "tests/*",
    ".venv/*",
]

[tool.coverage.report]
precision = 2
exclude_also = [
    "if TYPE_CHECKING:",
    "if __name__ == '__main__':",
    "pragma: no cover",
    "raise NotImplementedError",
]

[tool.codespell]
skip = ".git,.venv,uv.lock"
check-filenames = true
check-hidden = true
ignore-words-list = "crate,fo,nd"

[tool.vulture]
paths = ["."]
min_confidence = 80
exclude = [".venv", "tests"]

[tool.wily]
path = "."
extensions = [".py"]
exclude = [".venv", "tests"]
archiver = "git"


[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportMissingImports = "information"
reportMissingTypeStubs = "none"
include = ["."]
exclude = ["**/.venv/**", "**/tests/**"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
